Anadida la funcionalidad de listar por ID, (sin manejo de excepciones)

+++ src/main/java/com/sofkau/crudPersona/servicios/InterfazServicioPersona.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface InterfazServicioPersona {
 
     public List<Persona> Listar();
-    public Persona listarId(long id);
+    public Persona listarId(int id);
     public Persona guardar(Persona persona);
     public void borrar(int id);
     public Persona actualizar(Persona persona);
